Add three new functions and clean some code
Add the function to_json(), retweet(), ratio_of_likes(). Cleaned a bit
the code. The code implements the scheduler but is commented out to leave
the first real call to the function when more features are added to the
project. The check_recent_tweets() and search_for_users() shouldn´t
be used anymore - or at least in the foreseable future.